refactor: split test_utils.py into modular structure
- Move FakeModbusClient, FakeModbusClientPerformance, FakeModbusResponse to tests/fakes/modbus.py
- Extract demo register data to tests/fakes/responses.py
- Create pytest fixtures in tests/fixtures/ (modbus, homeassistant, registers)
- Move module loading helpers to tests/helpers/modules.py
- Move assertion helpers to tests/helpers/assertions.py
- Update all test imports to use new module structure
- Remove test_utils.py God Object
